修改

+++ application/api/controller/user/UserFeedback.php
@@ -40,14 +40,18 @@ class UserFeedback extends  Father
             $input['pic']=str_replace($qian, '', $input['pic']);    //清除前面传过来的数据里面的空格 换行
             $new_pic=explode(',',$input['pic']);//用户筛选过的图片
             $count=count($pic_data);
-            for($i=0;$i<$count;$i++){
-                $path_pic=FileHelper::helper()->getWebsitePath($pic_data[$i]);
-                //删除用户删除的图片
-                if (!in_array($path_pic,$new_pic)){
-                    FileHelper::helper()->unlink($pic_data[$i]);
-                    unset($pic_data[$i]);
+            //如果是ios传的就不去循环了  会出现超自然现象
+            if (!isset($input['type'])){
+                for($i=0;$i<$count;$i++){
+                    $path_pic=FileHelper::helper()->getWebsitePath($pic_data[$i]);
+                    //删除用户删除的图片
+                    if (!in_array($path_pic,$new_pic)){
+                        FileHelper::helper()->unlink($pic_data[$i]);
+                        unset($pic_data[$i]);
+                    }
                 }
             }
+
             $input['pic']=implode(',',$pic_data);
             unset($input['token']);
         }
